desafio-24-06-2025/pyproject.toml
[project]
name = "desafio-24-06-2025"
version = "0.1.0"
description = "Challenge with CrewAI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.121.1",
    "litellm>=1.68.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
